Alert for confirming deletions

+++ fe/js/post.js
@@ -124,7 +124,11 @@ let createPostButtons = (data) => {
         deleteButton.appendChild(deleteIcon);
         deleteButton.addEventListener('click', (e) => {
             e.preventDefault()
-            deletePost(data.id);
+            confirmWarning('Delete post', 'Are you sure you want to delete this post?').then((value) => {
+              if(value.isConfirmed){
+                deletePost(data.id);
+              }
+            });
         });
 
     }
